client/requirements.txt
psycopg2
pyyaml